changed VAOnePole class to use doubles

+++ WPKorg35/VAOnePoleFilter.cpp
@@ -19,28 +19,28 @@ CVAOnePoleFilter::~CVAOnePoleFilter(void)
 // recalc coeffs -- NOTE: not used for Korg35 Filter
 void CVAOnePoleFilter::updateFilter()
 {
-	float wd = 2*pi*m_fFc;          
-	float T  = 1/m_fSampleRate;             
-	float wa = (2/T)*tan(wd*T/2); 
-	float g  = wa*T/2;            
+	double wd = 2*pi*m_fFc;          
+	double T  = 1/m_fSampleRate;             
+	double wa = (2/T)*tan(wd*T/2); 
+	double g  = wa*T/2;            
 
 	m_fAlpha = g/(1.0 + g);
 }
 
 // do the filter
-float CVAOnePoleFilter::doFilter(float xn)
+double CVAOnePoleFilter::doFilter(double xn)
 {
 	// calculate v(n)
-	float vn = (xn - m_fZ1)*m_fAlpha;
+	double vn = (xn - m_fZ1)*m_fAlpha;
 
 	// form LP output
-	float lpf = vn + m_fZ1;
+	double lpf = vn + m_fZ1;
 
 	// update memory
 	m_fZ1 = vn + lpf;
 
 	// do the HPF
-	float hpf = xn - lpf;
+	double hpf = xn - lpf;
 
 	if(m_uFilterType == LPF1)
 		return lpf;
